fix moved messages
